feat!: rename the supabase package to supabase_dart
BREAKING CHANGE: The pure Dart client is published as supabase_dart instead
of supabase, and its library entrypoint is supabase_dart.dart. Depend on
supabase_dart and import package:supabase_dart/supabase_dart.dart.
supabase_flutter users are unaffected.
