+++ .github/workflows/release-pana.yml
@@ -27,9 +27,9 @@ jobs:
         package:
           - iceberg
           - postgrest
-          - supabase
           - supabase_auth
           - supabase_common
+          - supabase_dart
           - supabase_flutter
           - supabase_functions
           - supabase_realtime
