+++ .github/workflows/test.yml
@@ -64,7 +64,7 @@ jobs:
           echo "Affected packages:"
           echo "$CHANGED"
 
-          DART_PACKAGES="iceberg postgrest supabase supabase_auth supabase_common supabase_functions supabase_realtime supabase_storage supabase_typegen yet_another_json_isolate"
+          DART_PACKAGES="iceberg postgrest supabase_auth supabase_common supabase_dart supabase_functions supabase_realtime supabase_storage supabase_typegen yet_another_json_isolate"
 
           entries=()
           for package in $DART_PACKAGES; do
@@ -72,11 +72,11 @@ jobs:
               continue
             fi
             case "$package" in
-              postgrest|supabase|supabase_auth|supabase_realtime|supabase_storage) backend=true ;;
+              postgrest|supabase_auth|supabase_dart|supabase_realtime|supabase_storage) backend=true ;;
               *) backend=false ;;
             esac
             case "$package" in
-              postgrest|supabase|supabase_auth|supabase_realtime|supabase_storage) concurrency="1" ;;
+              postgrest|supabase_auth|supabase_dart|supabase_realtime|supabase_storage) concurrency="1" ;;
               *) concurrency="0" ;;
             esac
             entries+=("{\"name\":\"$package\",\"backend\":$backend,\"concurrency\":\"$concurrency\"}")
@@ -344,7 +344,7 @@ jobs:
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           parallel-finished: true
-          carryforward: 'iceberg,postgrest,supabase,supabase_auth,supabase_common,supabase_functions,supabase_realtime,supabase_storage,supabase_typegen,yet_another_json_isolate,supabase_flutter'
+          carryforward: 'iceberg,postgrest,supabase_auth,supabase_common,supabase_dart,supabase_functions,supabase_realtime,supabase_storage,supabase_typegen,yet_another_json_isolate,supabase_flutter'
           fail-on-error: false
 
   test-result:
