+++ packages/supabase_dart/example/.gitignore
