+++ packages/supabase_dart/example/analysis_options.yaml
