+++ packages/supabase_dart/example/pubspec.yaml
@@ -1,4 +1,4 @@
-name: supabase_example
+name: supabase_dart_example
 description: An absolute bare-bones supabase web app.
 version: 1.0.0
 publish_to: none
@@ -10,7 +10,7 @@ resolution: workspace
 
 dependencies:
   web: '>=1.0.0 <2.0.0'
-  supabase: ^2.16.0
+  supabase_dart: ^2.16.0
 
 dev_dependencies:
   build_runner: any
