+++ packages/supabase_dart/lib/src/realtime_client_options.dart
