+++ packages/supabase_dart/lib/src/supabase_constants.dart
@@ -1,4 +1,4 @@
-import 'package:supabase/src/version.dart';
+import 'package:supabase_dart/src/version.dart';
 import 'package:supabase_common/supabase_common.dart';
 import 'package:meta/meta.dart';
