+++ packages/supabase_dart/lib/src/supabase_query_schema.dart
@@ -1,5 +1,5 @@
 import 'package:http/http.dart';
-import 'package:supabase/supabase.dart';
+import 'package:supabase_dart/supabase_dart.dart';
 import 'package:yet_another_json_isolate/yet_another_json_isolate.dart';
 
 import 'counter.dart';
