+++ packages/supabase_dart/lib/src/version.dart
