+++ packages/supabase_dart/pubspec.yaml
@@ -1,8 +1,8 @@
-name: supabase
+name: supabase_dart
 description: A dart client for Supabase. This client makes it simple for developers to build secure and scalable products.
 version: 2.16.0
 homepage: 'https://supabase.com'
-repository: 'https://github.com/supabase/supabase-flutter/tree/main/packages/supabase'
+repository: 'https://github.com/supabase/supabase-flutter/tree/main/packages/supabase_dart'
 issue_tracker: 'https://github.com/supabase/supabase-flutter/issues'
 documentation: 'https://supabase.com/docs/reference/dart/introduction'
 topics:
