+++ packages/supabase_dart/test/client_test.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:supabase/src/supabase_client.dart' as real;
-import 'package:supabase/supabase.dart' hide SupabaseClient;
+import 'package:supabase_dart/src/supabase_client.dart' as real;
+import 'package:supabase_dart/supabase_dart.dart' hide SupabaseClient;
 import 'package:supabase_common/supabase_common.dart';
 import 'package:test/test.dart';
 import 'package:yet_another_json_isolate/yet_another_json_isolate.dart';
