+++ packages/supabase_flutter/pubspec.yaml
@@ -26,8 +26,8 @@ dependencies:
   http: ^1.6.0
   meta: ^1.16.0
   passkeys_platform_interface: ^2.8.0
-  supabase: 2.16.0
   supabase_common: 0.1.2
+  supabase_dart: 2.16.0
   url_launcher: ^6.3.2
   shared_preferences: ^2.5.5
   logging: ^1.3.0
