+++ sdk-compliance.yaml
@@ -15,8 +15,8 @@
 #                           explains why.
 #
 # Verified against the source in this repository (packages/supabase_auth,
-# postgrest, supabase_storage, supabase_realtime, supabase_functions, supabase,
-# supabase_flutter).
+# postgrest, supabase_storage, supabase_realtime, supabase_functions,
+# supabase_dart, supabase_flutter).
 # Unlisted features default to not_implemented.
 
 sdk: flutter
